fix(product): keep context-doc resolution out of the login entry chunk
The outgoing-prompt serializer moves to its own module so
file-mention-draft-model (in the entry chunk via the chat input store)
no longer pulls the context-doc mention domain into the login
first-load. Main sits 30 gzip bytes under the login JS cap; the
in-model import blew it by 415 (CI: JS 502415/502000 B). Only the three
send sites, all in the lazy workspace chunk, import the resolver.
